functions/requirements.txt
Flask
openai
google-search-results
markupsafe
python-dotenv
flask-cors